requirements.txt
# Web framework and API
flask==3.0.1
flask-cors==5.0.1
flask-socketio==5.5.1

# Werkzeug version compatible with Flask 2.0.1
werkzeug==3.1.3

# Scientific and computer vision
numpy==1.26.4
opencv-python==4.11.0.86
filterpy==1.4.5
lap==0.5.12
scipy==1.15.2
matplotlib==3.10.1
pillow==11.1.0

# MongoDB
pymongo==4.13.1


# Geometry
shapely==2.0.7

# Deep learning
torch==1.12.0
ultralytics== 8.3.122

# Utilities
requests==2.28.2

# For GUI (if running locally, not in Docker)
#Pillow==8.4.0

# If you use tkinter, it's included with Python, but on Linux you may need:
# RUN apt-get install -y python3-tk 